# backend/src/firstdue/api/routes/incidents.py
"""The incident API: open, brief, stream, resolve, approve, log, close.

The stream is the part with a real contract. Server-sent events, ordered by
brief version, and **every frame is in the incident log before it is sent** --
``require_persisted()`` raises otherwise, so the ordering is enforced by the type
rather than by this module remembering to await something first.

Reconnect is by ``Last-Event-ID``, which SSE gives us for free: a tablet that
lost signal reconnects, sends the last version it saw, and gets everything after
it in order. The replay reads the same emissions from the same log, so what a
reconnecting commander sees is what the original stream sent -- not a fresh
render that might differ.
"""

from __future__ import annotations

import json
import time
from collections.abc import AsyncIterator
from datetime import datetime
from typing import Annotated, Any

from fastapi import APIRouter, Depends, Header, Query, status
from pydantic import BaseModel, ConfigDict, Field
from sse_starlette.sse import EventSourceResponse

from firstdue.api.dependencies import (
    Caller,
    require_profile_write,
    require_read,
    require_referral_write,
)
from firstdue.api.routes.health import get_container
from firstdue.container import Container
from firstdue.domain.briefs import BriefEmission
from firstdue.domain.enums import BenchmarkType
from firstdue.errors import NotFoundError, ValidationError
from firstdue.incident.controller import IncidentController
from firstdue.incident.fusion import ThermalFrame
from firstdue.incident.session import get_session, sessions
from firstdue.observability.context import get_correlation_id
from firstdue.observability.logging import get_logger
from firstdue.observability.metrics import METRICS

logger = get_logger(__name__)

router = APIRouter(tags=["incident"])


# ------------------------------------------------------------------- models


class OpenIncidentRequest(BaseModel):
    model_config = ConfigDict(extra="forbid")

    #: Whatever CAD sent: a display address or an address id. The city adapter
    #: normalises it; this endpoint does not guess.
    address: str = Field(min_length=1, max_length=200)
    cad_ref: str = Field(min_length=1, max_length=120)
    alarm_level: int = Field(default=1, ge=1, le=5)
    dispatched_at: datetime | None = None
    responding_agency_id: str = Field(default="sffd", max_length=120)
    mutual_aid_agreement_id: str | None = Field(default=None, max_length=120)


class OpenIncidentResponse(BaseModel):
    model_config = ConfigDict(extra="forbid")

    incident_id: str
    address_id: str
    profile_snapshot_id: str
    grant_id: str
    cold_start: bool
    dispatched_at: datetime
    elapsed_seconds: float
    #: The instant brief, already persisted. Version 1, no model call.
    brief: dict[str, Any]
    instant_brief_ms: float
    event_id: str


class ResolutionRequest(BaseModel):
    """What an IC saw during the 360."""

    model_config = ConfigDict(extra="forbid")

    conflict_id: str = Field(min_length=1, max_length=120)
    observed_value: str = Field(min_length=1, max_length=200)
    resolved_by: str = Field(min_length=1, max_length=120)
    note: str = Field(default="", max_length=500)


class ThermalFrameRequest(BaseModel):
    model_config = ConfigDict(extra="forbid")

    face: str = Field(min_length=1, max_length=20)
    region_temps_c: list[float] = Field(min_length=1, max_length=64)
    coverage: float = Field(default=1.0, ge=0.0, le=1.0)
    source: str = Field(default="recorded", max_length=60)


class ResourceRequestBody(BaseModel):
    model_config = ConfigDict(extra="forbid")

    kind_id: str = Field(min_length=1, max_length=60)
    detail: str = Field(default="", max_length=300)
    #: Present when a human has already approved this exact request.
    approval_id: str | None = Field(default=None, max_length=120)


class CloseIncidentRequest(BaseModel):
    model_config = ConfigDict(extra="forbid")

    closed_by: str = Field(min_length=1, max_length=120)


def _controller(container: Container) -> IncidentController:
    return get_session(container).controller


# --------------------------------------------------------------------- open


@router.post(
    "/incidents",
    response_model=OpenIncidentResponse,
    status_code=status.HTTP_201_CREATED,
    summary="Open an incident from a CAD dispatch",
)
async def open_incident(
    request: OpenIncidentRequest,
    container: Annotated[Container, Depends(get_container)],
    caller: Annotated[Caller, Depends(require_profile_write)],
) -> OpenIncidentResponse:
    """Open an incident and return the instant brief.

    The brief in this response is already in the incident log -- it is persisted
    before it is returned, exactly as it is before it is streamed.
    """
    session = get_session(container)
    opened = await session.controller.open(
        address=request.address,
        cad_ref=request.cad_ref,
        alarm_level=request.alarm_level,
        dispatched_at=request.dispatched_at,
        responding_agency_id=request.responding_agency_id,
        mutual_aid_agreement_id=request.mutual_aid_agreement_id,
        # The caller's id, when it sent one. Without this the incident mints a
        # fresh correlation id and the request that opened it can no longer be
        # joined to the audit trail or the trace -- which is the one question
        # anyone asks afterwards.
        correlation_id=get_correlation_id(),
    )

    started = time.perf_counter()
    emission = await session.emit_instant(opened)
    elapsed_ms = (time.perf_counter() - started) * 1000.0
    METRICS.record_time_to_first_line(elapsed_ms)

    if elapsed_ms > container.settings.instant_brief_budget_ms:
        # A defect, not a slow day. Logged as such.
        logger.error(
            "instant_brief_budget_exceeded",
            extra={
                "incident_id": opened.incident.incident_id,
                "elapsed_ms": round(elapsed_ms, 3),
                "budget_ms": container.settings.instant_brief_budget_ms,
            },
        )

    return OpenIncidentResponse(
        incident_id=opened.incident.incident_id,
        address_id=opened.incident.address_id,
        profile_snapshot_id=opened.snapshot_id,
        grant_id=opened.grant.grant_id,
        cold_start=opened.cold_start,
        dispatched_at=opened.incident.dispatched_at,
        elapsed_seconds=session.controller.elapsed_seconds(opened.incident),
        brief=emission.model_dump(mode="json"),
        instant_brief_ms=round(elapsed_ms, 3),
        event_id=opened.event_id,
    )


# -------------------------------------------------------------------- brief


@router.get(
    "/incidents/{incident_id}/brief",
    summary="The latest brief version",
)
async def latest_brief(
    incident_id: str,
    container: Annotated[Container, Depends(get_container)],
    caller: Annotated[Caller, Depends(require_read)],
) -> dict[str, Any]:
    session = get_session(container)
    emission = session.latest(incident_id)
    if emission is None:
        raise NotFoundError("no brief for this incident", details={"incident_id": incident_id})
    return emission.model_dump(mode="json")


@router.post(
    "/incidents/{incident_id}/brief/enrich",
    summary="Produce the enriched brief version",
)
async def enrich_brief(
    incident_id: str,
    container: Annotated[Container, Depends(get_container)],
    caller: Annotated[Caller, Depends(require_profile_write)],
) -> dict[str, Any]:
    """Add the model-composed stage.

    A model that is unavailable or returns something the contract rejects leaves
    the deterministic brief exactly as it was, and the emission says the
    narrative is unavailable.
    """
    session = get_session(container)
    started = time.perf_counter()
    emission = await session.emit_enriched(incident_id)
    METRICS.record_enriched_latency((time.perf_counter() - started) * 1000.0)
    return emission.model_dump(mode="json")


def _frame(emission: BriefEmission) -> dict[str, str]:
    """One SSE frame. The version is the event id, which is what resume uses."""
    return {
        "event": "brief",
        "id": str(emission.version),
        "data": json.dumps(emission.model_dump(mode="json"), sort_keys=True),
    }


@router.get(
    "/incidents/{incident_id}/stream",
    summary="Stream brief versions over SSE",
)
async def stream_brief(
    incident_id: str,
    container: Annotated[Container, Depends(get_container)],
    caller: Annotated[Caller, Depends(require_read)],
    last_event_id: Annotated[str | None, Header(alias="Last-Event-ID")] = None,
    after_version: Annotated[int | None, Query(ge=0)] = None,
) -> EventSourceResponse:
    """Stream every brief version, in order, resumable.

    A reconnecting tablet sends ``Last-Event-ID`` and receives everything after
    it. The frames come from the emissions the log already holds, so a resumed
    stream shows what the original one sent rather than a fresh render.

    Every frame calls ``require_persisted()`` before it is yielded. An emission
    that somehow reached this point unpersisted raises rather than being shown.
    """
    session = get_session(container)
    resume_from = _resume_point(last_event_id, after_version)

    async def frames() -> AsyncIterator[dict[str, str]]:
        for emission in session.emissions_after(incident_id, resume_from):
            # The gate. Nothing reaches a commander that is not in the record.
            yield _frame(emission.require_persisted())

    return EventSourceResponse(frames())


def _resume_point(last_event_id: str | None, after_version: int | None) -> int:
    """Where to resume from. The header wins; the query is for testing."""
    if last_event_id:
        try:
            return int(last_event_id)
        except ValueError:
            # A client sending a malformed resume point gets the whole stream
            # rather than an error: showing the brief again is always safe.
            return 0
    return after_version or 0


# --------------------------------------------------------------- the 360


@router.post(
    "/incidents/{incident_id}/resolutions",
    status_code=status.HTTP_201_CREATED,
    summary="Record an IC resolution during the 360",
)
async def resolve_conflict(
    incident_id: str,
    request: ResolutionRequest,
    container: Annotated[Container, Depends(get_container)],
    caller: Annotated[Caller, Depends(require_profile_write)],
) -> dict[str, Any]:
    """Settle a disagreement by looking at the building.

    An IC walking the 360 is a human observation, so it can close a conflict
    that no newer document could. It writes a live-observation fact, resolves
    the conflict, bumps ``profile_version``, and produces a marked amendment.
    """
    session = get_session(container)
    result = await session.resolve(
        incident_id,
        conflict_id=request.conflict_id,
        observed_value=request.observed_value,
        resolved_by=request.resolved_by,
        note=request.note,
    )
    return result


@router.post(
    "/incidents/{incident_id}/thermal",
    status_code=status.HTTP_202_ACCEPTED,
    summary="Register a thermal frame to a face",
)
async def register_thermal(
    incident_id: str,
    request: ThermalFrameRequest,
    container: Annotated[Container, Depends(get_container)],
    # A write, not a read: registering a frame amends the brief and appends to
    # the incident log. Reading geometry is not permission to change it.
    caller: Annotated[Caller, Depends(require_profile_write)],
) -> dict[str, Any]:
    """Register recorded or synthetic thermal footage to one face.

    Never presented as a live flight. Faces with no current frame stay
    UNSCANNED, and every reading carries the caveat that thermal imaging
    measures surface temperature and cannot see through walls.
    """
    from firstdue.domain.enums import FaceLabel

    try:
        face = FaceLabel(request.face.upper())
    except ValueError as exc:
        raise ValidationError("unknown face label", details={"face": request.face[:20]}) from exc

    frame = ThermalFrame(
        frame_id=container.ids.new_id("frame"),
        incident_id=incident_id,
        face=face,
        observed_at=container.clock.now(),
        region_temps_c=tuple(request.region_temps_c),
        coverage=request.coverage,
        source=request.source,
    )
    session = get_session(container)
    return await session.register_thermal(incident_id, frame)


@router.post(
    "/incidents/{incident_id}/resources",
    summary="Request a resource -- notification or commitment",
)
async def request_resource(
    incident_id: str,
    request: ResourceRequestBody,
    container: Annotated[Container, Depends(get_container)],
    caller: Annotated[Caller, Depends(require_referral_write)],
) -> dict[str, Any]:
    """Ask for a resource. The gateway decides whether a human must approve.

    Notifications go out autonomously. Anything that commits another agency's
    resources comes back ``REQUIRE_APPROVAL`` with a staged, prefilled card --
    and that distinction is made by policy, not by this endpoint.
    """
    session = get_session(container)
    outcome = await session.request_resource(
        incident_id,
        kind_id=request.kind_id,
        detail=request.detail,
        approval_id=request.approval_id,
    )
    return outcome.model_dump(mode="json")


@router.post(
    "/incidents/{incident_id}/approvals/{approval_id}",
    summary="Approve a staged action",
)
async def approve_action(
    incident_id: str,
    approval_id: str,
    container: Annotated[Container, Depends(get_container)],
    caller: Annotated[Caller, Depends(require_referral_write)],
) -> dict[str, Any]:
    """The human tap. Records who approved it, and then executes."""
    session = get_session(container)
    return await session.approve(incident_id, approval_id, decided_by=caller.subject)


# ----------------------------------------------------------------- the log


@router.get(
    "/incidents/{incident_id}/log",
    summary="The append-only incident log",
)
async def incident_log(
    incident_id: str,
    container: Annotated[Container, Depends(get_container)],
    caller: Annotated[Caller, Depends(require_read)],
) -> dict[str, Any]:
    """Everything recorded during the incident, in order, with content hashes."""
    log = await container.incident_log.get_log(incident_id)
    return {
        "incident_id": incident_id,
        "sealed_at": log.sealed_at.isoformat() if log.sealed_at else None,
        "entries": [
            {
                "sequence": entry.sequence,
                "entry_type": str(entry.entry_type),
                "occurred_at": entry.occurred_at.isoformat(),
                "profile_snapshot_id": entry.profile_snapshot_id,
                "content_hash": entry.content_hash,
                "written_to_rms_at": (
                    entry.written_to_rms_at.isoformat() if entry.written_to_rms_at else None
                ),
                "content": entry.content,
            }
            for entry in log.entries
        ],
        "unflushed": len(log.unflushed),
    }


@router.post(
    "/incidents/{incident_id}/benchmarks/{benchmark_type}",
    status_code=status.HTTP_201_CREATED,
    summary="Timestamp an operational benchmark",
)
async def record_benchmark(
    incident_id: str,
    benchmark_type: BenchmarkType,
    container: Annotated[Container, Depends(get_container)],
    caller: Annotated[Caller, Depends(require_profile_write)],
) -> dict[str, Any]:
    """Record that something happened, when it happened. Clerical, not tactical."""
    mark = await _controller(container).record_benchmark(
        incident_id, benchmark_type, recorded_by=caller.subject
    )
    return mark.model_dump(mode="json")


# ---------------------------------------------------------------- closing


@router.post(
    "/incidents/{incident_id}/close",
    summary="Close the incident",
)
async def close_incident(
    incident_id: str,
    request: CloseIncidentRequest,
    container: Annotated[Container, Depends(get_container)],
    caller: Annotated[Caller, Depends(require_profile_write)],
) -> dict[str, Any]:
    """Close: revoke the grant, seal the log, draft the report.

    Both the revocation and the seal happen. Authority ends when the incident
    does, and a log that could still be appended to afterwards is a log nobody
    can rely on.
    """
    result = await _controller(container).close(incident_id, closed_by=request.closed_by)
    sessions(container).forget(incident_id)
    return result.model_dump(mode="json")