# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BenchmarkType": {
        "description": "Operational benchmarks timestamped as they occur. Clerical, not tactical.",
        "enum": [
          "DISPATCH",
          "ARRIVAL",
          "WATER_ON_FIRE",
          "PRIMARY_SEARCH_COMPLETE",
          "PAR",
          "INCIDENT_CLOSED"
        ],
        "title": "BenchmarkType",
        "type": "string"
      },
...
      "CloseIncidentRequest": {
        "additionalProperties": false,
        "properties": {
          "closed_by": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Closed By",
            "type": "string"
          }
        },
        "required": [
          "closed_by"
        ],
        "title": "CloseIncidentRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpenIncidentRequest": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Address",
            "type": "string"
          },
          "alarm_level": {
            "default": 1,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Alarm Level",
            "type": "integer"
          },
          "cad_ref": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Cad Ref",
            "type": "string"
          },
          "dispatched_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatched At"
          },
          "mutual_aid_agreement_id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mutual Aid Agreement Id"
          },
          "responding_agency_id": {
            "default": "sffd",
            "maxLength": 120,
            "title": "Responding Agency Id",
            "type": "string"
          }
        },
        "required": [
          "address",
          "cad_ref"
        ],
        "title": "OpenIncidentRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Municipal structural intelligence as an institutional agent fleet.\n\nA decision-support prototype. It delivers information and performs clerical\nexecution. It never issues tactical recommendations, evacuation orders, crew\nassignments, or fire-behaviour predictions.",
    "title": "FIRST DUE",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/incidents": {
      "post": {
        "description": "Open an incident and return the instant brief.\n\nThe brief in this response is already in the incident log -- it is persisted\nbefore it is returned, exactly as it is before it is streamed.",
        "operationId": "open_incident_api_v1_incidents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenIncidentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenIncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open an incident from a CAD dispatch",
        "tags": [
          "incident"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/approvals/{approval_id}": {
      "post": {
        "description": "The human tap. Records who approved it, and then executes.",
        "operationId": "approve_action_api_v1_incidents__incident_id__approvals__approval_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Action Api V1 Incidents  Incident Id  Approvals  Approval Id  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve a staged action",
        "tags": [
          "incident"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/benchmarks/{benchmark_type}": {
      "post": {
        "description": "Record that something happened, when it happened. Clerical, not tactical.",
        "operationId": "record_benchmark_api_v1_incidents__incident_id__benchmarks__benchmark_type__post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "benchmark_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BenchmarkType"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Record Benchmark Api V1 Incidents  Incident Id  Benchmarks  Benchmark Type  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Timestamp an operational benchmark",
        "tags": [
          "incident"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/brief": {
      "get": {
        "operationId": "latest_brief_api_v1_incidents__incident_id__brief_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Latest Brief Api V1 Incidents  Incident Id  Brief Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The latest brief version",
        "tags": [
          "incident"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/brief/enrich": {
      "post": {
        "description": "Add the model-composed stage.\n\nA model that is unavailable or returns something the contract rejects leaves\nthe deterministic brief exactly as it was, and the emission says the\nnarrative is unavailable.",
        "operationId": "enrich_brief_api_v1_incidents__incident_id__brief_enrich_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Enrich Brief Api V1 Incidents  Incident Id  Brief Enrich Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Produce the enriched brief version",
        "tags": [
          "incident"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/close": {
      "post": {
        "description": "Close: revoke the grant, seal the log, draft the report.\n\nBoth the revocation and the seal happen. Authority ends when the incident\ndoes, and a log that could still be appended to afterwards is a log nobody\ncan rely on.",
        "operationId": "close_incident_api_v1_incidents__incident_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseIncidentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Close Incident Api V1 Incidents  Incident Id  Close Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close the incident",
        "tags": [
          "incident"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/log": {
      "get": {
        "description": "Everything recorded during the incident, in order, with content hashes.",
        "operationId": "incident_log_api_v1_incidents__incident_id__log_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Incident Log Api V1 Incidents  Incident Id  Log Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The append-only incident log",
        "tags": [
          "incident"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/resolutions": {
      "post": {
        "description": "Settle a disagreement by looking at the building.\n\nAn IC walking the 360 is a human observation, so it can close a conflict\nthat no newer document could. It writes a live-observation fact, resolves\nthe conflict, bumps ``profile_version``, and produces a marked amendment.",
        "operationId": "resolve_conflict_api_v1_incidents__incident_id__resolutions_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resolve Conflict Api V1 Incidents  Incident Id  Resolutions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record an IC resolution during the 360",
        "tags": [
          "incident"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/resources": {
      "post": {
        "description": "Ask for a resource. The gateway decides whether a human must approve.\n\nNotifications go out autonomously. Anything that commits another agency's\nresources comes back ``REQUIRE_APPROVAL`` with a staged, prefilled card --\nand that distinction is made by policy, not by this endpoint.",
        "operationId": "request_resource_api_v1_incidents__incident_id__resources_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Request Resource Api V1 Incidents  Incident Id  Resources Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request a resource -- notification or commitment",
        "tags": [
          "incident"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/stream": {
      "get": {
        "description": "Stream every brief version, in order, resumable.\n\nA reconnecting tablet sends ``Last-Event-ID`` and receives everything after\nit. The frames come from the emissions the log already holds, so a resumed\nstream shows what the original one sent rather than a fresh render.\n\nEvery frame calls ``require_persisted()`` before it is yielded. An emission\nthat somehow reached this point unpersisted raises rather than being shown.",
        "operationId": "stream_brief_api_v1_incidents__incident_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Version"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream brief versions over SSE",
        "tags": [
          "incident"
        ]
      }
    },
...
  }
}